fix msw

+++ src/gui/Utils.cpp
@@ -10,7 +10,7 @@ int get_line_from_istream(char *buf, size_t len, void *context)
         return -1;
     }
     is->getline(buf, len);
-    return is->gcount();
+    return (int)is->gcount();
 }
 
 int put_line_to_ostream(const char *buf, size_t len, void *context)
@@ -18,5 +18,5 @@ int put_line_to_ostream(const char *buf, size_t len, void *context)
     std::ostream *os = static_cast<std::ostream *>(context);
     os->write(buf, len);
     os->put('\n');
-    return len;
+    return (int)len;
 }
